requirements.txt
albucore==0.0.23
albumentations==1.4.24
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
data==0.4
decorator==5.2.1
dill==0.4.1
dnspython==2.8.0
doclayout_yolo==0.0.4
einops==0.8.2
email-validator==2.3.0
entmax==1.3
fastapi==0.129.0
fastapi-cli==0.0.21
fastapi-cloud-cli==0.11.0
fastar==0.8.0
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
funcsigs==1.0.2
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
latex==0.7.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
munch==4.0.0
networkx==3.6.1
numpy==2.4.2
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==3.0.0
pillow==12.1.0
pix2tex==0.1.4
polars==1.38.0
polars-runtime-32==1.38.0
psutil==7.2.2
py-cpuinfo==9.0.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
PyPDF2==3.0.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
rich==14.3.2
rich-toolkit==0.19.4
rignore==0.7.6
safetensors==0.7.0
scipy==1.17.0
seaborn==0.13.2
sentry-sdk==2.52.0
setuptools==80.10.2
shapely==2.1.2
shellingham==1.5.4
shutilwhich==1.1.0
simsimd==6.5.12
six==1.17.0
starlette==0.52.1
stringzilla==4.6.0
sympy==1.14.0
tempdir==0.7.1
thop==0.1.1.post2209072238
timm==0.5.4
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.0.0
typer==0.23.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.11
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
x-transformers==0.15.0
